port stable IPAM namespace fix onto develop's refactored module paths
develop relocated the IPAM constants, isOfKind, and getObjectDetailsUrl
into a per-entity domain/model structure and centralized query-param
names into the QSP enum. stable's namespace-preservation fix targeted the
old paths, which develop deleted; the auto-merge kept stable's new call
sites while silently dropping the modules they imported.

Remap to develop's paths (verified 1:1, same values):
- IP_NAMESPACE_GENERIC -> @/entities/ipam/ip-namespaces/domain/model/ip-namespace
- isOfKind             -> @/entities/schema/domain/rules/is-of-kind
- IPAM_QSP.NAMESPACE   -> QSP.IPAM_NAMESPACE (@/shared/config/qsp)
- getObjectDetailsUrl  -> @/entities/nodes/object/ui/routing/object-urls

Port stable's orphaned getObjectDetailsUrl unit test to object-urls.test.ts
next to its new home, preserving coverage develop lacked.
